+++ src/app/elements/pedigree/pedigree.component.ts
@@ -30,10 +30,16 @@ export class PedigreeComponent implements OnInit {
   getPopoverContent(goat?: Goat) {
     if (!goat) return '';
     const datePipe = new DatePipe('en-US');
-    return `<div>
-        <span class="fw-bold">Born</span>: <span class="fw-light">${datePipe.transform(goat?.dateOfBirth, 'longDate')}</span>
-        <br>
-        <span class="fw-bold">ID</span>: <span class="fw-light">${goat?.normalizeId}</span>
-      </div>`;
+    const popover: string[] = [];
+    if (goat.dateOfBirth) {
+      popover.push(`<span class="fw-bold">Born</span>: <span class="fw-light">${datePipe.transform(goat.dateOfBirth, 'longDate')}</span>`);
+    }
+    if (goat.normalizeId) {
+      popover.push(`<span class="fw-bold">ID</span>: <span class="fw-light">${goat.normalizeId}</span>`);
+    }
+    if (goat.ownerAccount?.displayName) {
+      popover.push(`<span class="fw-bold">Owned By</span>: <span class="fw-light">${goat.ownerAccount.displayName}</span>`);
+    }
+    return `<div>${popover.join('<br>')}</div>`;
   }
 }