+++ src/app/services/goat/goat.service.ts
@@ -118,8 +118,9 @@ export type Goat = Partial<{
   animalTattoo: { tattoo?: string; tattooLocation?: { name?: string; }; }[];
   id: number;
   sex: 'Female' | 'Male';
-  damId?: number;
-  sireId?: number;
+  damId: number;
+  sireId: number;
+  ownerAccount: { displayName?: string; };
 }>;
 export const Goat = {
   nickname: 'Your Goats Farm Name',